# Conflicts:
#	src/main/java/com/mercadolibre/group8_bootcamp_finalproject/dtos/InboundOrderDTO.java
#	src/main/java/com/mercadolibre/group8_bootcamp_finalproject/services/InboundOrderServiceImpl.java
#	src/test/java/com/mercadolibre/group8_bootcamp_finalproject/unit/InboundOrderUnitTest.java

+++ src/main/java/com/mercadolibre/group8_bootcamp_finalproject/dtos/WarehouseSectionDTO.java
@@ -11,8 +11,8 @@ import javax.validation.constraints.NotNull;
 public class WarehouseSectionDTO {
 
     @NotNull(message = "section code is required")
-    private Integer sectionCode;
-
+    private Long sectionCode;
+// TODO não serve pra nada mas está no teste
     @NotNull(message = "warehouse code is required")
-    private Integer warehouseCode;
+    private Long warehouseCode;
 }
